Added support for * in select.
This cleans up how column names were handled as well. The parser is now
required to have a "columns" attribute that is a list of column names.

query.columns = list of selected columns or None for all (*)
parser.columns = query(parser).column_names = list of all columns in the source file